requirements.txt
aniso8601==9.0.1
blinker==1.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
opencv-python-headless==4.5.4.58
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
